requirements.txt
# Core dependencies
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2

# PDF processing
pdfplumber==0.10.3
tabula-py==2.9.0
PyPDF2==3.0.1
fpdf==1.7.2  # For generating sample PDFs
pymupdf==1.23.9  # For PDF visualization with annotations

# CSV and data processing
pandas==2.1.1
numpy==1.26.1

# Other utilities
python-dotenv==1.0.0
tqdm==4.66.1
faker==20.1.0  # For generating synthetic data 
